requirements.txt
# Baseline OCR tools
olmocr
pypdf
pillow

# Core AI stack (Pinned to avoid V5 conflict)
transformers>=4.45.0,<5.0.0
torch>=2.4.0
torchvision>=0.19.0
accelerate

# Mac-Specific (Only on Apple Silicon)
# Environment markers ensure these don't break Windows/Linux installs
mlx; sys_platform == 'darwin' and platform_machine == 'arm64'
mlx-vlm; sys_platform == 'darwin' and platform_machine == 'arm64'

# Common conflicts fixes
numpy<2.3.0
fsspec<=2025.10.0